+++ backend/authentication/admin.py
@@ -7,7 +7,7 @@ from django.contrib.auth.forms import ReadOnlyPasswordHashField
 from django.contrib.auth.models import Group
 from django.core.exceptions import ValidationError
 
-from .models import (
+from authentication.models import (
     Support,
     SupportEntityType,
     UserModel,
