+++ backend/authentication/serializers.py
@@ -10,7 +10,7 @@ from django.utils.translation import gettext as _
 from rest_framework import serializers
 from rest_framework.authtoken.models import Token
 
-from .models import UserModel
+from authentication.models import UserModel
 
 USER = get_user_model()
