+++ backend/authentication/urls.py
@@ -1,6 +1,6 @@
 from django.urls import path
 
-from . import views
+from authentication import views
 
 app_name = "authentication"
