+++ backend/authentication/views.py
@@ -15,8 +15,8 @@ from rest_framework.request import Request
 from rest_framework.response import Response
 from rest_framework.views import APIView
 
-from .models import UserModel
-from .serializers import (
+from authentication.models import UserModel
+from authentication.serializers import (
     DeleteUserResponseSerializer,
     LoginSerializer,
     PasswordResetSerializer,
