+++ backend/communities/groups/serializers.py
@@ -1,0 +1,70 @@
+"""
+Serializers for groups in the communities app.
+"""
+
+from typing import Any
+
+from rest_framework import serializers
+
+from communities.groups.models import (
+    Group,
+    GroupImage,
+    GroupMember,
+    GroupText,
+)
+from content.serializers import LocationSerializer, ResourceSerializer
+from events.serializers import EventSerializer
+
+# MARK: Main Tables
+
+
+class GroupTextSerializer(serializers.ModelSerializer[GroupText]):
+    class Meta:
+        model = GroupText
+        fields = "__all__"
+
+
+class GroupSerializer(serializers.ModelSerializer[Group]):
+    texts = GroupTextSerializer(many=True, read_only=True)
+    location = LocationSerializer(read_only=True)
+    resources = ResourceSerializer(many=True, read_only=True)
+    events = EventSerializer(many=True, read_only=True)
+
+    class Meta:
+        model = Group
+        extra_kwargs = {
+            "created_by": {"read_only": True},
+        }
+
+        fields = "__all__"
+
+    def validate(self, data: dict[str, Any]) -> dict[str, Any]:
+        if data.get("terms_checked") is False:
+            raise serializers.ValidationError(
+                "You must accept the terms of service to create a group."
+            )
+
+        return data
+
+    def create(self, validated_data: dict[str, Any]) -> Group:
+        group = Group.objects.create(**validated_data)
+
+        if group:
+            GroupText.objects.create(group=group)
+
+        return group
+
+
+# MARK: Bridge Tables
+
+
+class GroupImageSerializer(serializers.ModelSerializer[GroupImage]):
+    class Meta:
+        model = GroupImage
+        fields = "__all__"
+
+
+class GroupMemberSerializer(serializers.ModelSerializer[GroupMember]):
+    class Meta:
+        model = GroupMember
+        fields = "__all__"
