+++ backend/communities/urls.py
@@ -1,7 +1,9 @@
 from django.urls import include, path
 from rest_framework.routers import DefaultRouter
 
-from . import views
+from communities.groups.views import GroupViewSet
+from communities.organizations.views import OrganizationViewSet
+from communities.views import StatusViewSet
 
 app_name = "communities"
 
@@ -9,11 +11,11 @@ router = DefaultRouter()
 
 # MARK: Main Tables
 
-router.register(prefix=r"groups", viewset=views.GroupViewSet)
+router.register(prefix=r"groups", viewset=GroupViewSet)
 router.register(
-    prefix=r"organizations", viewset=views.OrganizationViewSet, basename="organization"
+    prefix=r"organizations", viewset=OrganizationViewSet, basename="organization"
 )
-router.register(prefix=r"statuses", viewset=views.StatusViewSet)
+router.register(prefix=r"statuses", viewset=StatusViewSet)
 
 urlpatterns = [
     path("", include(router.urls)),
