+++ backend/content/admin.py
@@ -1,6 +1,6 @@
 from django.contrib import admin
 
-from .models import (
+from content.models import (
     Discussion,
     DiscussionEntry,
     Faq,
