+++ backend/core/urls.py
@@ -3,7 +3,10 @@ URL configuration for the activist backend.
 
 The `urlpatterns` list routes URLs to views. For more information please see:
     https://docs.djangoproject.com/en/4.2/topics/http/urls/
-Examples:
+
+Examples
+--------
+
 Function views
     1. Add an import:  from my_app import views
     2. Add a URL to urlpatterns:  path('', views.home, name='home')
