+++ frontend/composables/useLinkURL.ts
@@ -17,8 +17,8 @@ export function useLinkURL(props: {
     let url: string = "";
     if (props.organization) {
       url = `/organizations/${props.organization.id}`;
-      // } else if (props.group) {
-      // url = `/organizations/${props.group.organization.id}/groups/${props.group.id}`;
+    } else if (props.group) {
+      url = `/organizations/${props.group.parentOrgId}/groups/${props.group.id}`;
     } else if (props.event) {
       url = `/events/${props.event.id}`;
     } else if (props.resource) {
